# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VEO Asset Management API",
        "description": "REST API for VEO Asset Management System mobile application",
        "contact": {
            "email": "support@veoset.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://veoset.test",
            "description": "VEO Asset Management API Server"
        }
    ],
    "paths": {
...
        "/api/maintenance/tasks": {
            "get": {
                "tags": [
                    "Maintenance"
                ],
                "summary": "Get maintenance tasks",
                "operationId": "250f5023a0d00fa0102ec4982afa65c4",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_progress",
                                "completed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filter by priority",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high",
                                "emergency"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Maintenance tasks list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Maintenance"
                ],
                "summary": "Create maintenance task",
                "operationId": "9948b48f836e029e1bc28ca5aed15b35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "asset_id",
                                    "title",
                                    "priority"
                                ],
                                "properties": {
                                    "asset_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high",
                                            "emergency"
                                        ]
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/maintenance/tasks/{id}": {
            "get": {
                "tags": [
                    "Maintenance"
                ],
                "summary": "Get maintenance task details",
                "operationId": "f5ece2f13e649c6963e4c4a20cd192da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/maintenance/tasks/{id}/status": {
            "put": {
                "tags": [
                    "Maintenance"
                ],
                "summary": "Update task status",
                "operationId": "d639bbd0dcf2f4753bbed94e61bce437",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "completed",
                                            "cancelled"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string"
                                    },
                                    "completion_notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task status updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/maintenance/schedule": {
            "get": {
                "tags": [
                    "Maintenance"
                ],
                "summary": "Get maintenance schedule",
                "operationId": "0ba9fc427b060c32ea895b2e807d5fb0",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Specific date (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Maintenance schedule"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}